// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const OwnerType = {
  user: 'user',
  tenant: 'tenant'
} as const

export type OwnerType = (typeof OwnerType)[keyof typeof OwnerType]


export const TypePaymentState = {
  PENDIENTE: 'PENDIENTE',
  RECHAZADO: 'RECHAZADO',
  ANULADO: 'ANULADO',
  FINALIZADO: 'FINALIZADO'
} as const

export type TypePaymentState = (typeof TypePaymentState)[keyof typeof TypePaymentState]


export const TypeDocument = {
  dni: 'dni',
  passport: 'passport',
  driver_license: 'driver_license',
  other: 'other'
} as const

export type TypeDocument = (typeof TypeDocument)[keyof typeof TypeDocument]


export const Gender = {
  male: 'male',
  female: 'female',
  other: 'other'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const HabitType = {
  general: 'general',
  do_more: 'do_more',
  do_less: 'do_less'
} as const

export type HabitType = (typeof HabitType)[keyof typeof HabitType]


export const ProgressStatus = {
  done: 'done',
  not_done: 'not_done'
} as const

export type ProgressStatus = (typeof ProgressStatus)[keyof typeof ProgressStatus]